backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pgvector
langchain
openai
langchain_community
langchain_postgres
tiktoken